src/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.5
load-dotenv==0.1.0
MarkupSafe==3.0.2
python-dotenv==1.0.1
Werkzeug==3.1.3
WTForms==3.2.1
